feat(context): introduce Browser property
closes #939
