+++ src/PlaywrightSharp/Browser.cs
@@ -140,7 +140,6 @@ namespace PlaywrightSharp
         {
             var context = (await Channel.NewContextAsync(options ?? new BrowserContextOptions()).ConfigureAwait(false)).Object;
             BrowserContextsList.Add(context);
-            context.Browser = this;
             return context;
         }
