+++ src/PlaywrightSharp/Transport/Protocol/PageInitializer.cs
@@ -7,5 +7,7 @@ namespace PlaywrightSharp.Transport.Protocol
         public FrameChannel MainFrame { get; set; }
 
         public ViewportSize ViewportSize { get; set; }
+
+        public bool IsClosed { get; set; }
     }
 }
